Create USC-10 Sec-253
Presidential declaration to the military to disperse government officials committing crimes against the United States public after I ran for president in 2020 had both political parties cheat to include electronically and physically attacking me multiple times during and after the election, which was exactly the kind of thing that I ran for. Criminals across the board and this has all been reported to the United Nations and three letter agencies for decades for some of the paperwork and especially in the last five years, so there is no excuse and it was done on purpose because I had been reporting them.  You don’t get a choice about what happens to you.